+++ src/index.js
@@ -17,6 +17,7 @@ export { modInverse } from "./arithmetic/modInverse.js";
 export { sign } from "./arithmetic/sign.js";
 
 // Combinatorics Functions
+export { bell } from "./combinatorics/bell.js";
 export { combination } from "./combinatorics/combination.js";
 export { factorial } from "./combinatorics/factorial.js";
 export { fibonacci } from "./combinatorics/fibonacci.js";